Pass a stable request id along to help with dup detection server side
For the case where a request makes it to us we process it but perhaps the network between fails to communicate back to the client that we did and the client retries. If we know the request id, we can try to avoid duplicates with some sort of light coordination.