Remove event receiver stuff from flipper
I am not sure this makes sense here and it will cause some friction in iterating on events for cloud. I'd rather focus on cloud events and if others think it would be useful in OSS then I can always extract it out of cloud and back into the gem.